requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver-binary==122.0.6261.94.0
click==8.1.7
consonance==0.1.2
cryptography==42.0.8
dissononce==0.34.3
dnspython==2.6.1
email_validator==2.2.0
evpn==1.0.3
fake-useragent==1.5.1
Flask==3.0.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.1.0
MarkupSafe==2.1.5
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==23.2
pillow==10.4.0
playwright==1.45.1
pluggy==1.4.0
protobuf==3.20.0
psutil==5.9.8
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyee==11.1.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyQt6==6.7.1
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.7.2
pyqt6-tools==6.4.2.3.3
PyQt6_sip==13.8.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.0.2
pytest-base-url==2.1.0
pytest-playwright==0.4.4
python-axolotl==0.2.2
python-axolotl-curve25519==0.4.1.post2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-xlib==0.15
pytweening==1.2.0
pywhatkit==5.4
pywhatsapp==0.0.7
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
readline==6.2.4.1
requests==2.31.0
selenium==4.23.1
sip==6.8.6
six==1.10.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
text-unidecode==1.3
transitions==0.9.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.8.0
Werkzeug==3.0.4
wikipedia==1.4.0
wsproto==1.2.0
yowsup2==2.5.0
